requirements.txt
aiogram>=3.4.0,<4
python-dotenv>=1.0.0
python-docx>=1.1.0
openpyxl>=3.1.2
APScheduler>=3.10.4
pytest>=8.0.0
